Change the condition for `ploneglossary.css`.
The new condition does not give an error when accessing the Dexterity
control panel, which does not allow access to ``meta_type``.
Unfortunately, a `hasattr` does not help in this case.

Note that if you have customized PloneGlossary and kept the same
meta_type but have given it a different portal_type, you will need to
change the condition accordingly.  Includes upgrade step.

This fixes #3.