Support highlighting text from fields of dexterity items.
We now use the indexer to get the SearchableText, which works for both
Archetypes and dexterity.

Until now for dexterity only terms in the title and description would
be highlighted, because that is what is returned by SearchableText by
default.  If those terms occured in a rich text field too, they would
be highlighted.

We recommend to not enable highlighting on files.